hw-android-2-04 Работа над ошибками 1. List<String> urlList - теперь static поле 2. class RetorfitUtil перенесен в MainActivity и обращается через  MainActivity.this
